% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diallel_pedigree.R
\name{diallel_pedigree}
\alias{diallel_pedigree}
\title{Build Diallel Pedigree}
\usage{
diallel_pedigree(
  parents = NULL,
  popsize = NULL,
  subpopsize = NULL,
  selfs = 0,
  padsize = 4,
  nextinteger = FALSE
)
}
\arguments{
\item{parents}{number of parents.}

\item{popsize}{population size.}

\item{subpopsize}{subpopulation size.}

\item{selfs}{number of selfed populations}

\item{padsize}{string length for the individual name number}

\item{nextinteger}{if TRUE all the subpopulations will have same size, total population size will changed accordingly to the next possible integer, if FALSE sub populations will be unbalenced}
}
\value{
half-diallel pedigree.
}
\description{
Build half-diallel pedigree based on number of parents, number of selfs, and population size
}
\examples{
ped <- diallel_pedigree(parents=5,popsize=1000,selfs=2)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
