% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_haplo.R
\name{fake_haplo}
\alias{fake_haplo}
\title{Build fake haplotypes}
\usage{
fake_haplo(n = NULL, m = NULL, seed = NULL)
}
\arguments{
\item{n}{number of haplotypes (integer).}

\item{m}{number of markers (integer).}

\item{seed}{seed to be used in the sampling}
}
\value{
Matrix with the Relationship between the individuals.
}
\description{
Build fake haplotypes based on given number of markers and number of haplotypes
}
\examples{
haps <- fake_haplo(n=10,m=100,seed=12345)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
