% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_pedigree.R
\name{linear_pedigree}
\alias{linear_pedigree}
\title{Build Linear Pedigree}
\usage{
linear_pedigree(parents = 4, popsize = 100, selfs = 1, padsize = 4)
}
\arguments{
\item{parents}{number of parents.}

\item{popsize}{population size.}

\item{selfs}{number of selfed populations}

\item{padsize}{string length for the individual name number}
}
\value{
linear pedigree.
}
\description{
Build linear pedigree based on number of parents, number of selfs, and population size
}
\examples{
ped <- linear_pedigree(parents=4,popsize=100,selfs=1)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
