% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreesimR.R
\name{pedigreesimR}
\alias{pedigreesimR}
\title{PedigreeSim Management}
\usage{
pedigreesimR(map, haplotypes, pedigree, centromere = NULL,
  prefPairing = 0, quadrivalents = 0, ploidy = 2,
  workingfolder = "PedigreeSimR_files", mapfunction = "HALDANE",
  chromosome = "A", seed = NULL, allownochiasmata = 1,
  naturalpairing = 1, parallelquadrivalents = 0,
  pairedcentromeres = 0, mapwidthpad = 4)
}
\arguments{
\item{map}{vector of the marker position.}

\item{haplotypes}{matrix with haplotype information to be draw}

\item{pedigree}{pedigree data frame}

\item{centromere}{numeric with its position}

\item{prefPairing}{numeric}

\item{quadrivalents}{numeric}

\item{ploidy}{integer even number}

\item{workingfolder}{string with folder name to write input/output files}

\item{mapfunction}{"HALDANE" or "KOSAMBI"}

\item{chromosome}{string}

\item{seed}{integer to be used to sample haplotypes}

\item{allownochiasmata}{numeric}

\item{naturalpairing}{numeric}

\item{parallelquadrivalents}{numeric}

\item{pairedcentromeres}{numeric}

\item{mapwidthpad}{numeric length of marker code}
}
\value{
nothing?
}
\description{
Wrap-up function to run PedigreeSim software
}
\examples{
map = 1:100
haplotypes = fake_haplo(100,100,seed=1234)
pedigree = diallel_pedigree(parents=7,popsize=1000,selfs=2)
pedigreesimR(map,haplotypes,pedigree)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
