% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTLsim.R
\name{QTLsim}
\alias{QTLsim}
\title{Simulate QTL + residual effects}
\usage{
QTLsim(
  parents = 3,
  ploidy = 4,
  workingfolder = "PedigreeSimR_files",
  QTLmarker = NULL,
  QTLh2 = 0.3,
  run_diaQTL = FALSE
)
}
\arguments{
\item{parents}{number of parents in the pedigree}

\item{ploidy}{integer even number}

\item{workingfolder}{string with folder name to load the input and write output files}

\item{QTLmarker}{name of marker to be used as QTL, if NULL a marker is sampled randomly (default)}

\item{QTLh2}{numeric between 0 and 1 with the QTL heritability}

\item{run_diaQTL}{logical, if TRUE runs diaQTL scan1 and fitQTL functions in the data.}
}
\value{
if run_diaQTL=TRUE, return a list with diaQTL results. Otherwise, returns NULL.
}
\description{
Simulates phenotypic data on top of the pedigreeSimR output and save then as diaQTL csv format
}
\examples{
\dontrun{
map = 1:100
haplotypes = fake_haplo(m=50,n=100,seed=1234)
pedigree = diallel_pedigree(parents=3,popsize=100)
pedigreesimR(map,haplotypes,pedigree,workingfolder="PedigreeSimR_files")
QTLsim(parents=3,
       ploidy=4,
       workingfolder="PedigreeSimR_files",
       QTLmarker=NULL,
       QTLh2=0,
       run_diaQTL=FALSE)
}
}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
