% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_pedigree.R
\name{round_pedigree}
\alias{round_pedigree}
\title{Build Round-Robin Pedigree}
\usage{
round_pedigree(parents = NULL, popsize = NULL, subpopsize = NULL,
  padsize = 4, nextinteger = FALSE)
}
\arguments{
\item{parents}{number of parents.}

\item{popsize}{population size.}

\item{subpopsize}{subpopulation size.}

\item{nextinteger}{if TRUE all the subpopulations will have same size, total population size will changed accordingly to the next possible integer, if FALSE sub populations will be unbalenced}

\item{padize}{string length for the individual name number}
}
\value{
single round-robin pedigree.
}
\description{
Build single round-robin pedigree based on number of parents and population size
}
\examples{
haps <- fake_haplo(n=10,map=100,seed=12345)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
