% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_pedigree.R
\name{self_pedigree}
\alias{self_pedigree}
\title{Build Self Pedigree}
\usage{
self_pedigree(parents = NULL, popsize = NULL, padsize = 4, nextinteger = FALSE)
}
\arguments{
\item{parents}{number of parents.}

\item{popsize}{population size.}

\item{padsize}{string length for the individual name number}

\item{nextinteger}{if TRUE all the subpopulations will have same size, total population size will changed accordingly to the next possible integer, if FALSE sub populations will be unbalenced}
}
\value{
half-diallel pedigree.
}
\description{
Build selfs pedigree based on number of parents and population size
}
\examples{
ped <- self_pedigree(parents=5,popsize=1000)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
