% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_pedigree.R
\name{star_pedigree}
\alias{star_pedigree}
\title{Build Star Pedigree}
\usage{
star_pedigree(parents = 4, popsize = 100, selfs = 1, padsize = 4)
}
\arguments{
\item{parents}{number of parents.}

\item{popsize}{population size.}

\item{selfs}{seed to be used in the sampling}

\item{padize}{string length for the individual name number}

\item{nextinteger}{if TRUE all the subpopulations will have same size, total population size will changed accordingly to the next possible integer, if FALSE sub populations will be unbalenced}
}
\value{
half-diallel pedigree.
}
\description{
Build star pedigree based on number of parents, number of selfs, and population size
}
\examples{
ped <- star_pedigree(parents=4,popsize=100,selfs=1)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
