% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_pedigree.R
\name{round_pedigree}
\alias{round_pedigree}
\title{Build Round-Robin Pedigree}
\usage{
round_pedigree(
  parents = NULL,
  popsize = NULL,
  subpopsize = NULL,
  padsize = 4,
  nextinteger = FALSE,
  selfs = 0
)
}
\arguments{
\item{parents}{number of parents.}

\item{popsize}{population size.}

\item{subpopsize}{subpopulation size.}

\item{padsize}{string length for the individual name number}

\item{nextinteger}{if TRUE all the subpopulations will have same size, total population size will changed accordingly to the next possible integer, if FALSE sub populations will be unbalenced}

\item{selfs}{number of selfed populations}
}
\value{
single round-robin pedigree.
}
\description{
Build single round-robin pedigree based on number of parents and population size
}
\examples{
haps <- round_pedigree(parents=3,popsize=200)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
