% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_pedigree.R
\name{star_pedigree}
\alias{star_pedigree}
\title{Build Star Pedigree}
\usage{
star_pedigree(parents = 4, popsize = 100, selfs = 1, padsize = 4)
}
\arguments{
\item{parents}{number of parents.}

\item{popsize}{population size.}

\item{selfs}{number of selfed populations}

\item{padsize}{string length for the individual name number}
}
\value{
star pedigree.
}
\description{
Build star pedigree based on number of parents, number of selfs, and population size
}
\examples{
ped <- star_pedigree(parents=4,popsize=100,selfs=1)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
